Document upload check for Referral reply added
